requirements.txt
SQLAlchemy==2.0.0b1
fastapi==0.103.2
python-dotenv==1.0.0
pydantic==2.4.2
uvicorn==0.23.2
fastapi-mail==1.4.1
email-validator==2.0.0.post2
pydantic==2.4.2
starlette==0.27.0
